% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessEnvBias.R
\name{assessEnvBias}
\alias{assessEnvBias}
\title{\code{assessEnvBias}}
\usage{
assessEnvBias(dat, envCols, periods, ...)
}
\arguments{
\item{dat}{string. A data.frame containing columns for species name (NA if not identified), an identifier (usually taxonomic group name),
and spatial uncertainty.}

\item{periods}{String. A list of time periods. For example, for two periods, the first spanning 1950 to 1990, and the second 1991 to 2019: periods = list(1950:1990, 1991:2019).}

\item{envDat}{String. A raster::stack object with n layers, one for each environmental variable to be considered.}
}
\value{
A list with two elements if filter = FALSE and three elements if filter = TRUE. The elements are 1) data (summary of spatial uncertainty),
2) a ggplot object and 3) the input data with the user-defined spatialUncertainty filter applied.
}
\description{
This function plots the distribution of your data in environmental space. To do this it uses a principal component analysis to reduce your environmental data to two dimensions for visualization.
}
