% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessRepeatVisits.R
\name{assessRepeatVisits}
\alias{assessRepeatVisits}
\title{\code{assessRepeatVisits}}
\usage{
assessRepeatVisits(dat, periods, spatRes)
}
\arguments{
\item{dat}{String. Data.frame with columns for lat, lon, and identifier (usually taxonomic group name).}

\item{periods}{String. A list of time periods. For example, for two periods, the first spanning 1950 to 1990, and the second 1991 to 2019: periods = list(1950:1990, 1991:2019).}

\item{spatRes}{Numeric. Spatial resolution at which to calculate repeat visits. Depends on CRS (use metres for eastings and northings, and decimals for lat lon).}
}
\description{
This function calculates the number of repeat visits across cells of a user-defined spatial resolution. Repeats are divided
by the number of years in each user-defined period to give the average number of repeat samples per year on each grid cell.
}
