% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessSpeciesBias.R
\name{assessSpeciesBias}
\alias{assessSpeciesBias}
\title{\code{assessSpeciesBias}}
\usage{
assessSpeciesBias(dat, periods, res)
}
\arguments{
\item{dat}{string. A data.frame containing columns for species name (NA if not identified), an identifier (usually taxonomic group name),
and year (NA if not known).}

\item{periods}{String. A list of time periods. For example, for two periods, the first spanning 1950 to 1990, and the second 1991 to 2019: periods = list(1950:1990, 1991:2019).}

\item{res}{Logical. Spatial resolution at which to grid the data for estimation of range size.}
}
\description{
This function assesses the proportionality of species observed range sizes and number of records. For each period, the function regresses
the number of grid cells with records (proxy for range size) on the number of records. It uses the r2 value from these regressions
as an index proportionality between the two measures. A lower r2 indicates lower proportionality between range size and the number of records relative
to a higher score.
}
\examples{

}
