% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessSpeciesID.R
\name{assessSpeciesID}
\alias{assessSpeciesID}
\title{\code{assessSpeciesID}}
\usage{
assessSpeciesID(dat, periods, type, maxSpatUncertainty = NULL)
}
\arguments{
\item{dat}{string. A data.frame containing columns species (species name), x (x coordinate), y (y coordinate), year, spatialUncertainty and identifier.}

\item{periods}{String. A list of time periods. For example, for two periods, the first spanning 1950 to 1990, and the second 1991 to 2019: periods = list(1950:1990, 1991:2019).}

\item{type}{String. "proportion" or "count".}

\item{maxSpatUncertainty}{Numeric. Maximum permitted spatial uncertainty. All records more uncertain than this value will be dropped. Units must match the units in your data.}
}
\value{
A list with two elements: a ggplot2 object and the data underpinning the plot.
}
\description{
This function calculates the proportion (or counts) of records identified to species level over time..
}
