% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessSpeciesNumber.R
\name{assessSpeciesNumber}
\alias{assessSpeciesNumber}
\title{\code{assessSpeciesNumber}}
\usage{
assessSpeciesNumber(dat, periods)
}
\arguments{
\item{dat}{string. A data.frame containing columns for species name (NA if not identified), an identifier (usually taxonomic group name),
and year (NA if not known).}

\item{periods}{String. A list of time periods. For example, for two periods, the first spanning 1950 to 1990, and the second 1991 to 2019: periods = list(1950:1990, 1991:2019).}
}
\description{
This function calculates the number of species recorded in each year.
}
\seealso{
\code{\link{assessSpeciesID}} which gives the number of species identified to species level.
}
