% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backgroundEnvDat.R
\docType{data}
\name{backgroundEnvDat}
\alias{backgroundEnvDat}
\title{Background environmental data for the UK}
\format{
An object of class \code{"data.frame"}
}
\source{
\href{10.1002/joc.5086}{doi}
}
\usage{
data(backgroundEnvDat)
}
\description{
Environmental data from worldclim layers (worldclim.org) at 4000 random points in the UK.
}
\examples{
data(backgroundEnvDat)
}
\references{
Fick and Hijmans. (2017) International Journal of Climatology.
(\href{10.1002/joc.5086}{doi})
}
\keyword{datasets}
