% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callunaData.R
\docType{data}
\name{callunaData}
\alias{callunaData}
\title{Arabidopsis QTL data on gravitropism}
\format{
An object of class \code{"data.frame"}.
}
\usage{
data(callunaData)
}
\description{
Data on Calluna vulgaris occupancy in Britain. Each row represents 1 1x1 km grid square. There are columns
for sample inclusion in two time-periods (1=sampled and =0 not sampled), occupancy (1=occupued and 0= not occupued) and several
auxiliary variables.
}
\references{
Boyd et al. (2023) EcoEvoRxiv https://doi.org/10.32942/X2359P
}
\keyword{datasets}
