% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random40Species.R
\docType{data}
\name{random40Species}
\alias{random40Species}
\title{Simulated occurrence data.}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(random40Species)
}
\description{
Simulated occurrence data for 40 species in the UK in the format required by occAssess. The data are simulated to have random distributions in space and time.
Data prevalence varies among species by a factr of 5 with a mean of ~ 1000 records. All records are resolved to species level.
}
\examples{
data(random40Species)
}
\keyword{datasets}
