% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random40SpeciesEnvDat.R
\docType{data}
\name{random40SpeciesEnvDat}
\alias{random40SpeciesEnvDat}
\title{Background environmental data at the locations of the simulated data random40Species.}
\format{
An object of class \code{"data.frame"}
}
\source{
\href{10.1002/joc.5086}{doi}
}
\usage{
data(random40SpeciesEnvDat)
}
\description{
Environmental data from worldclim layers (worldclim.org) at the locations of the simulated data..
}
\examples{
data(random40SpeciesEnvDat)
}
\references{
Fick and Hijmans. (2017) International Journal of Climatology.
(\href{10.1002/joc.5086}{doi})
}
\keyword{datasets}
