% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDat.R
\docType{data}
\name{simDat}
\alias{simDat}
\title{Simulated occurrence data.}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(simDat)
}
\description{
Simulate occurrence data for 11 species in the UK. The data are drawn from a weighted random sample across the UK, where
the weights are equal to predicted species' distributions predicted in another project. For each simulated species, the
weights are taken from an anonymous actual species' predicted distributions. The data are generated uniformly over the period
2001 to 2010.
}
\examples{
data(simDat)
}
\keyword{datasets}
