% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simEnvDat.R
\docType{data}
\name{simEnvDat}
\alias{simEnvDat}
\title{Environmental data at the locations of the simulated data simDat.}
\format{
An object of class \code{"data.frame"}
}
\source{
\href{10.1002/joc.5086}{doi}
}
\usage{
data(simEnvDat)
}
\description{
Environmental data from worldclim layers (worldclim.org) at the locations of the simulated data simDat.
}
\examples{
data(simEnvDat)
}
\references{
Fick and Hijmans. (2017) International Journal of Climatology.
(\href{10.1002/joc.5086}{doi})
}
\keyword{datasets}
