% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKWGS84mask.R
\docType{data}
\name{UKWGS84mask}
\alias{UKWGS84mask}
\title{Example mask layer for assessSpatialBias(). .}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(UKWGS84mask)
}
\description{
Raster layer defining UK with finite values and elsewhere as NA.
}
\examples{
data(UKWGS84mask)
}
\keyword{datasets}
