% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessRarityBias.R
\name{assessRarityBias}
\alias{assessRarityBias}
\title{\code{assessRarityBias}}
\usage{
assessRarityBias(
  dat,
  species,
  x,
  y,
  year,
  spatialUncertainty,
  identifier,
  periods,
  res,
  prevPerPeriod,
  maxSpatUncertainty = NULL,
  metric = "r2"
)
}
\arguments{
\item{dat}{string. A data.frame containing columns for species name, x coordinates, y coordinates, spatialUncertainty, year and an identifier (used to group the data - heuristic will be calculated for each group).}

\item{species}{Character string. column name in dat giving species names.}

\item{x}{string. Column name in dat giving x coordinates. Any coordinate and spatial reference systems are permitted.}

\item{y}{string. Column name in dat giving y coordinates. Any coordinate and spatial reference systems are permitted.}

\item{year}{string. Column name in dat giving years.}

\item{spatialUncertainty}{String. Column name in dat giving uncertainty associated with x and y. Any units are permitted.}

\item{identifier}{String. Column name in dat giving record "identifiers". Identifiers are used to group the data; heuristics will be calculated separately for each group.}

\item{periods}{String. A list of time periods. For example, for two periods, the first spanning 1950 to 1990, and the second 1991 to 2019: periods = list(1950:1990, 1991:2019).}

\item{res}{Numeric. Spatial resolution at which to grid the data for estimation of range size.}

\item{prevPerPeriod}{Logical. If TRUE then prevalence (number of grid cells with data) is calculated per period and if FALSE then prevalence is calculated over all periods.}

\item{maxSpatUncertainty}{Numeric. Maximum permitted spatial uncertainty. All records more uncertain than this value will be dropped. Units must match the units in your data.}

\item{metric}{String. One of "r2" (coefficient of variation) or "cor" (Pearson's correlation coefficient). Two ways of summarising the linear regressions of sample size on recorded range size.
Note that, for large sample size, use of "cor" is more computationally demanding than "r2".}
}
\value{
A list with two elements: a ggplot2 object and the data underpinning the plot.
}
\description{
This function assesses the proportionality of species observed range sizes and number of records. For each period, the function regresses
the number of grid cells with records (proxy for range size) on the number of records. It uses the r2 value from these regressions
as an index proportionality between the two measures. A lower r2 indicates lower proportionality between range size and the number of records relative
to a higher score. Note that an index will not be calculated for any period/ identifier combination with fewer than 5 species. In these cases the regressions on which the indices are based are likely meaningless.
}
