% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessSpatialCov.R
\name{assessSpatialCov}
\alias{assessSpatialCov}
\title{\code{assessSpatialCov}}
\usage{
assessSpatialCov(dat, periods, res, logCount = FALSE, countries)
}
\arguments{
\item{dat}{string. A data.frame containing columns for species name (NA if not identified), an identifier (usually taxonomic group name),
and year (NA if not known).}

\item{periods}{String. A list of time periods. For example, for two periods, the first spanning 1950 to 1990, and the second 1991 to 2019: periods = list(1950:1990, 1991:2019).}

\item{res}{Numeric. Spatial resolution at which to grid the occurrence data.}

\item{logCount}{Logical. Whether to log transform counts for visual purposes. Useful where there is large variation in counts across cells.}
}
\description{
This function grids occurrence data then maps it as counts per grid cell in geographic spece. .
}
\seealso{
\code{\link{assessSpeciesID}} which gives the number of species identified to species level.
}
