% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessSpatialCov.R
\name{assessSpatialCov}
\alias{assessSpatialCov}
\title{\code{assessSpatialCov}}
\usage{
assessSpatialCov(
  dat,
  res,
  logCount = FALSE,
  countries,
  shp = NULL,
  periods,
  maxSpatUncertainty = NULL
)
}
\arguments{
\item{dat}{string. A data.frame containing columns species (species name), x (x coordinate), y (y coordinate), year, spatialUncertainty and identifier.}

\item{res}{Numeric. Spatial resolution at which to grid the occurrence data.}

\item{logCount}{Logical. Whether to log transform counts for visual purposes. Useful where there is large variation in counts across cells.}

\item{countries}{String or vector. Country names to be passed to ggplot2::map_data. Needed to add borders to your plot. Countries should only be used if you are working on
the WGS84 coordinate reference system. Otherwise see shp.}

\item{shp}{String. If you are not working on WGS84, then you will need to provide a shapefile (spatialPolygons or spatialPolygonsDataFrame) with the country borders on the relevant crs for plotting.}

\item{periods}{String. A list of time periods. For example, for two periods, the first spanning 1950 to 1990, and the second 1991 to 2019: periods = list(1950:1990, 1991:2019).}

\item{maxSpatUncertainty}{Numeric. Maximum permitted spatial uncertainty. All records more uncertain than this value will be dropped. Units must match the units in your data.}
}
\value{
A list with n ggplot2 objects where n is the number of levels in the identifier field of dat.
}
\description{
This function grids occurrence data then maps it as counts per grid cell in geographic spece. .
}
\seealso{
\code{\link{assessSpatialBias}} which gives a measure of how far your data eviates from a random distribution in space.
}
