% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessSpeciesID.R
\name{assessSpeciesID}
\alias{assessSpeciesID}
\title{\code{assessSpeciesID}}
\usage{
assessSpeciesID(dat, periods, type)
}
\arguments{
\item{dat}{string. A data.frame containing columns for species name (NA if not identified), an identifier (usually taxonomic group name),
and year (NA if not known).}

\item{periods}{String. A list of time periods. For example, for two periods, the first spanning 1950 to 1990, and the second 1991 to 2019: periods = list(1950:1990, 1991:2019).}

\item{filter}{Logical. Whether or not to return the input data after filtering based on a threshold spatialUncertainty.}

\item{threshold}{Numeric. Threshold to be used if filter = TRUE.}
}
\value{
A list with two elements if filter = FALSE and three elements if filter = TRUE. The elements are 1) data (summary of proportions identified over time),
2) a ggplot object and 3) the input data with the user-defined spatialUncertainty filter applied.
}
\description{
This function calculates the proportion (or counts) of records identified to species level over time..
}
