% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessSpeciesNumber.R
\name{assessSpeciesNumber}
\alias{assessSpeciesNumber}
\title{\code{assessSpeciesNumber}}
\usage{
assessSpeciesNumber(
  dat,
  species,
  x,
  y,
  year,
  spatialUncertainty,
  identifier,
  periods,
  maxSpatUncertainty = NULL
)
}
\arguments{
\item{dat}{string. A data.frame containing columns for species name, x coordinates, y coordinates, spatialUncertainty, year and an identifier (used to group the data - heuristic will be calculated for each group).}

\item{species}{Character string. column name in dat giving species names.}

\item{x}{string. Column name in dat giving x coordinates. Any coordinate and spatial reference systems are permitted.}

\item{y}{string. Column name in dat giving y coordinates. Any coordinate and spatial reference systems are permitted.}

\item{year}{string. Column name in dat giving years.}

\item{spatialUncertainty}{String. Column name in dat giving uncertainty associated with x and y. Any units are permitted.}

\item{identifier}{String. Column name in dat giving record "identifiers". Identifiers are used to group the data; heuristics will be calculated separately for each group.}

\item{periods}{String. A list of time periods. For example, for two periods, the first spanning 1950 to 1990, and the second 1991 to 2019: periods = list(1950:1990, 1991:2019).}

\item{maxSpatUncertainty}{Numeric. Maximum permitted spatial uncertainty. All records more uncertain than this value will be dropped. Units must match the units in your data.}
}
\value{
A list with two elements: a ggplot2 object and the data underpinning the plot.
}
\description{
This function calculates the number of species recorded in each year.
}
\seealso{
\code{\link{assessSpeciesID}} which gives the number or proportion of records identified to species level.
}
